mazeBuilder integration
