--- html/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="rus">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="author" content="Maxim Neverovsky">
-    <meta name="description" content="Тестовая страница">
-    <meta name="keywords" content="тест,проверка,html">
-    <title>Задачи 6 модуль</title>
-   <link rel="stylesheet" href="/css/main.css">
- </head>
-<body>
-   <div id="some_block">+</div>
-   <div id="header">Шапка</div>
-   <ul id="nav">
-     <li>Главная</li>
-      <li>Про нас</li>
-      <li>Контакты</li>
-   </ul>
-   <div id="main">Основная часть</div>
-
-   <div id="aside">Боковая часть</div>
-   <div id="footer">Футер</div>
-</body>
-</html>
