--- html/new_index.html
@@ -1,15 +1,0 @@
- <!DOCTYPE html>
- <html lang="ru">
- <head>
-     <meta charset="UTF-8">
-     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-     <title>Document</title>
- </head>
- <body>
-    
-
-     <script src="/js/main.js">
-     </script>
- </body>
- </html>
