--- html/project.html
@@ -1,60 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-    <meta http-equiv="content-type" content="text/html;charset=windows-1251">
-    <style>
-        body {
-            background: silver
-        }
-
-        #myTBL td {
-            background: white;
-        }
-
-        #myTBL td.act {
-            background: black;
-        }
-    </style>
-    <script>
-        function myFunc(e)                               // в одних браузерах при клике сгенерируется объект event и передастся функции как аргумент e, в других - не передастся
-        {
-            var evt = window.event || e,                      // где event не передался, пользуем window.event
-                obj = evt.srcElement || evt.target;           // получаем тег, по которому кликнули - синтаксис для двух групп браузеров
-            if (obj.tagName == 'TABLE') return;               // кликнули по бордеру или по межячеечному пространству, т.е. не точно по ячейке 
-            while (obj.tagName != 'TD') obj = obj.parentNode; // в ячейке могут быть внутренние теги (см. b22) - вверх по иерархии доходим от них до тега TD
-            obj.className = obj.className ? '' : 'act';       // если фон был белым, станет чёрным, и - наоборот
-        }
-
-        // только ПОСЛЕ окончания загрузки браузер будет знать о существовании тега с id="myTBL"
-        // и только ТОГДА мы можем дать браузеру команду назначить функцию обработки события клика по этому тегу
-        onload = function () {
-            var trg = document.getElementById('myTBL');
-            if (document.addEventListener) trg.addEventListener('click', myFunc); // синтаксис назначения функции для одних браузеров
-            else if (document.attachEvent) trg.attachEvent('onclick', myFunc);    // для других
-            else trg.onclick = myFunc;                                             // для третьих
-        }
-    </script>
-</head>
-
-<body>
-    <table id="myTBL" cellpadding="10" border="1">
-        <tr>
-            <td></td>
-            <td></td>
-            <td></td>
-        </tr>
-        <tr>
-            <td></td>
-            <td></td>
-            <td></td>
-        </tr>
-        <tr>
-            <td></td>
-            <td></td>
-            <td></td>
-        </tr>
-    </table>
-</body>
-
-</html>
